Kill entire child process tree after timeout
Node's child.kill() method only kills the child process, not any of
its descendants. Making the child a detached process means that the
entire child process group can be killed (using the negative of the
pid). This is useful as the command line solver does seem to create
child processes that were not otherwise killed.